logrotate: split into a package and extract pure period math
logrotate.py was a 657-line module mixing a ~500-line rotating-file handler
with the LogRotationManager facade and a block of pure period/date
arithmetic. Split it into a gracenote2epg/logrotate/ package:

  periods.py  pure date math (next rollover, week start, period info,
              completeness) — no handler state, independently testable
  handler.py  the copytruncate handler (period methods now delegate to
              periods.py)
  manager.py  the LogRotationManager facade

Behaviour preserved: the extracted period functions were verified identical
to the pre-split methods across 7200 date/mode checks, and a functional
multi-period startup-rotation test passes.
