+++ gracenote2epg/logrotate/__init__.py
@@ -1,0 +1,6 @@
+"""gracenote2epg.logrotate - built-in log rotation package."""
+
+from .handler import CopyTruncateTimedRotatingFileHandler
+from .manager import LogRotationManager
+
+__all__ = ["CopyTruncateTimedRotatingFileHandler", "LogRotationManager"]
